- Now you can see the weather for next week - TodayWeatherView style changes - Weather icons are renamed to match API icons and unnecessary ones are removed - Other small fixes

+++ app/WeatherApp/Models/DailyWeatherModel.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct DailyWeatherModel: Codable {
-    let dt: Int64
+    let dt: Double
     let temp: DailyTemperatureModel
     let weather: [WeatherInfoModel]
